Fix: Response type serialization tests
